third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/worklet-import-blocked-expected.txt

CONSOLE ERROR: Loading the script 'http://127.0.0.1:8000/resources/worklet.js' violates the following Content Security Policy directive: "script-src 'unsafe-inline'". Note that 'script-src-elem' was not explicitly set, so 'script-src' is used as a fallback. The action has been blocked.
This is a testharness.js-based test.
Harness: the test ran to completion.
